--- resources/assets/js/api/api.js
@@ -1,6 +1,0 @@
-class Api {
-    constructor(client, routes) {
-        this.client = client;
-        this.routes = routes;
-    }
-}
