+++ resources/assets/js/api/index.js
@@ -1,26 +1,7 @@
-import axios from 'axios';
+import client from './client';
 
-import Auth from './resources/auth';
-import UserResource from './resources/user_resource';
-
-const config = {};
-const client = axios.create(config);
-
-// Add a request interceptor
-client.interceptors.request.use((config) => {
-    const token = Auth.token();
-    config.headers.Authorization = `Bearer ${token}`;
-    return config;
-}, (error) => {
-    return Promise.reject(error);
-});
-
-// Add a response interceptor
-client.interceptors.response.use((response) => {
-    return Promise.resolve(response);
-}, (error) => {
-    return Promise.reject(error);
-});
+import Auth from './endpoints/auth';
+import UserResource from './endpoints/users';
 
 const api = {
     auth: new Auth(client),
